alerts for alllllllllllllllllllllll ouh

+++ app/Http/Controllers/PackageController.php
@@ -1,16 +1,25 @@
 <?php
 
 namespace App\Http\Controllers;
-
+use App\Models\Tenantpackage;
 use App\Models\Packagedetails;
 use Illuminate\Http\Request;
 
 Class PackageController extends Controller
 {
     public function store(Request $request){
-        Packagedetails::create(['name' => $request->package,
-        'price' => $request->price,
-        'description' => $request->description]);
+        try {
+            Packagedetails::create([
+                'name' => $request->package,
+                'price' => $request->price,
+                'description' => $request->description
+            ]);
+    
+            session()->flash('success', 'The package has been successfully created.');
+        } catch (\Exception $e) {
+            session()->flash('failed', 'An error occurred while creating the package.');
+        }
+    
         return back();
     }
 
@@ -32,8 +41,17 @@ Class PackageController extends Controller
     public function destroy($id)
     {
         $Package = Packagedetails::findOrFail($id);
-        $Package->delete();
+        $specificColumnValue = $Package->name;
+        $Tenant = Tenantpackage::where('tenant_package', '=', $specificColumnValue)->get();
 
-        return redirect()->back()->with('success', 'Package deleted successfully.');
+
+        if ($Tenant->isEmpty()) {
+            $Package->delete();
+            session()->flash('success', 'The package has been successfully created.');
+            return redirect()->back();
+        } else {
+            session()->flash('Failed', 'There are organizations subscribed to this package.');
+            return redirect()->back();
+        }
     }
 }
